smg/requirement.txt
boto3==1.14.47
botocore==1.17.47
dnspython==2.0.0
docutils==0.15.2
jmespath==0.10.0
numpy==1.19.1
pandas==1.1.1
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
